Cover delete file edge cases and refactor

+++ api/service/project.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"log"
 	"net/http"
@@ -10,6 +11,7 @@ import (
 	"strings"
 
 	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	"github.com/thomasmendez/personal-website-backend/api/database"
 	"github.com/thomasmendez/personal-website-backend/api/models"
 )
@@ -270,14 +272,34 @@ func (s *Service) deleteProjectHandler(ctx context.Context, request events.APIGa
 	if existingProject.MediaLink != nil && strings.Contains(*existingProject.MediaLink, ".s3.amazonaws.com") {
 		// TODO: Verify file exists before attempting to delete
 		// If 404 not found, don't worry continue to delete from database
-		err = s.S3.DeleteFileFromS3(ctx, *existingProject.MediaLink)
-		if err != nil {
-			log.Printf("error in deleting file from S3: %v", err)
-			return events.APIGatewayProxyResponse{
-				StatusCode: http.StatusInternalServerError,
-				Body:       resError(http.StatusInternalServerError),
-			}, err
+
+		// Check if file exists
+		// exists, err := s.S3.FileExistsInS3(ctx, *existingProject.MediaLink)
+		// if err != nil {
+		// 	log.Printf("error in getting file from S3: %v", err)
+		// }
+
+		if exists, err := s.S3.FileExistsInS3(ctx, *existingProject.MediaLink); exists {
+			err = s.S3.DeleteFileFromS3(ctx, *existingProject.MediaLink)
+			if err != nil {
+				log.Printf("error in deleting file from S3: %v", err)
+				return events.APIGatewayProxyResponse{
+					StatusCode: http.StatusInternalServerError,
+					Body:       resError(http.StatusInternalServerError),
+				}, err
+			}
+		} else {
+			if err != nil {
+				// Check if it's a "not found" error
+				var nf *types.NoSuchKey
+				if errors.As(err, &nf) {
+					return false, nil // File doesn't exist, but no error
+				}
+				return false, err // Some other error occurred
+			}
+			log.Printf("file does not exist in S3")
 		}
+
 	}
 
 	log.Printf("deleting project: %v", deleteProject)
